# Much faster conversion from Skia to Pillow