--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -29,9 +29,4 @@ jobs:
       run: poetry install
       if: steps.cached-dependencies.outputs.cache-hit != 'true'
     - name: Test
-      run: poetry run pytest tests --cov=./ --cov-report=xml
-#    - name: "Upload coverage to Codecov"
-#      uses: codecov/codecov-action@v2
-#      with:
-#        fail_ci_if_error: true
-#        token: ${{ secrets.CODECOV_TOKEN }}
+      run: poetry run pytest tests