--- a/pyproject.toml
+++ b/pyproject.toml
@@ -13,7 +13,7 @@ skia-python = {version = "^121.0b6", allow-prereleases = true}
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.2.1"
 pytest-watch = "^4.2.0"
-pytest-cov = "^3.0.0"
+pytest-cov = "^5.0.0"
 pylint = "^3.0.2"
 pycodestyle = "^2.8.0"
 Sphinx = "^5.2.3"