--- a/pytamaro/io.py
+++ b/pytamaro/io.py
@@ -15,7 +15,7 @@ from typing import List
 from PIL import Image as PILImageMod
 from PIL.Image import Image as PILImage
 from skia import (Canvas, FILEWStream, Image, Rect, Surface, SVGCanvas, kPNG,
-                  kUnpremul_AlphaType)
+                  kRGBA_8888_ColorType, kUnpremul_AlphaType)
 
 from pytamaro.checks import check_graphic, check_type
 from pytamaro.debug import add_debug_info
@@ -101,7 +101,9 @@ def graphic_to_pillow_image(graphic: Graphic) -> PILImage:
     :param graphic: graphic to be rendered and converted
     :returns: rendered graphic as a Pillow image
     """
-    return PILImageMod.fromarray(graphic_to_image(graphic).convert(alphaType=kUnpremul_AlphaType))
+    return PILImageMod.fromarray(graphic_to_image(graphic).convert(
+        alphaType=kUnpremul_AlphaType, colorType=kRGBA_8888_ColorType)
+    )
 
 
 # pylint: disable-next=invalid-name