Pick a strategy that does not make an expensive copy of the buffer each time ([list of strategies](https://github.com/kyamagu/skia-python/blob/e0b030c14e33f70880cfcc502eaeed898a77fc3c/notebooks/Python-Image-IO.ipynb)).

This process has to be done for each frame in animations.

Benchmark: on my machine, rendering 2K simple frames goes from 10.9s to 5.8s. The remaining slow part is encoding the GIF, done by Pillow.

```
Benchmark 1: python3 before.py
  Time (mean ± σ):     10.935 s ±  0.120 s    [User: 11.045 s, System: 2.041 s]
  Range (min … max):   10.807 s … 11.113 s    5 runs
Benchmark 1: python3 after.py
  Time (mean ± σ):      5.814 s ±  0.013 s    [User: 6.042 s, System: 2.118 s]
  Range (min … max):    5.800 s …  5.830 s    5 runs
```