@lucach 2024-06-03
Tests were succeeding on my machine but failing on the CI because it turns out that [Skia's default color type for images is platform-dependent](https://github.com/kyamagu/skia-python/pull/250). On Linux (therefore, also with the standard GitHub Runners), the default is `kBGRA_8888_ColorType` (BGR instead of RGB, respecting the endianness).